# api-service/internal/items/http/app/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  github_com_abdulazizax_mini-twitter_api-service_genproto_user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  github_com_abdulazizax_mini-twitter_api-service_genproto_user.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  github_com_abdulazizax_mini-twitter_api-service_genproto_user.RawResponse:
    properties:
      message:
        type: string
    type: object
  github_com_abdulazizax_mini-twitter_api-service_genproto_user.RegisterUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  github_com_abdulazizax_mini-twitter_api-service_genproto_user.SendVerificationCodeRequest:
    properties:
      email:
        type: string
    type: object
  github_com_abdulazizax_mini-twitter_api-service_genproto_user.UpdateUserPasswordRequest:
    properties:
      email:
        type: string
      new_password:
        type: string
      verification_code:
        type: integer
    type: object
  github_com_abdulazizax_mini-twitter_api-service_genproto_user.User:
    properties:
      bio:
        type: string
      created_at:
        $ref: '#/definitions/timestamppb.Timestamp'
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      profile_picture_url:
        type: string
      updated_at:
        $ref: '#/definitions/timestamppb.Timestamp'
      username:
        type: string
    type: object
...
  github_com_abdulazizax_mini-twitter_api-service_internal_models_user-handler.UpdateUserRequest:
    properties:
      bio:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      profile_picture_url:
        type: string
    type: object
  timestamppb.Timestamp:
    properties:
      nanos:
        description: |-
          Non-negative fractions of a second at nanosecond resolution. Negative
          second values with fractions must still have non-negative nanos values
          that count forward in time. Must be from 0 to 999,999,999
          inclusive.
        type: integer
      seconds:
        description: |-
          Represents seconds of UTC time since Unix epoch
          1970-01-01T00:00:00Z. Must be from 0001-01-01T00:00:00Z to
          9999-12-31T23:59:59Z inclusive.
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Endpoints for MiniTwitter
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: '# MiniTwitter'
  version: 1.03.67.83.145
paths:
  /auth/send-verification-email:
    post:
      consumes:
      - application/json
      description: Send a verification code to the user's email
      parameters:
      - description: Send verification code request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.SendVerificationCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.RawResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Send verification code
      tags:
      - Auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return login information
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
      summary: User login
      tags:
      - Auth
  /auth/sign-out:
    post:
      description: Log out the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: User logout
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Register a new user with the provided information
      parameters:
      - description: User registration information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Register a new user
      tags:
      - Auth
  /auth/update-password:
    post:
      consumes:
      - application/json
      description: Update the user's password
      parameters:
      - description: Update user password request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.UpdateUserPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Update user password
      tags:
      - Auth
  /user:
    get:
      description: Get the information of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_genproto_user.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Get user information
      tags:
      - Users
    post:
      description: Delete the authenticated user's account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the information of the authenticated user
      parameters:
      - description: User update information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_abdulazizax_mini-twitter_api-service_internal_models_user-handler.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      security:
      - BearerAuth: []
      summary: Update user information
      tags:
      - Users
...
  /user/uploadmedia:
    post:
      consumes:
      - multipart/form-data
      description: Upload a media file
      parameters:
      - description: UploadMediaForm
        in: formData
        name: file
        required: true
        type: file
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema: {}
      security:
      - BearerAuth: []
      summary: uploadFile
      tags:
      - Users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"